backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
gunicorn==22.0.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
groq==0.11.0
python-dotenv==1.0.1
